replace StaffController with StaffRestController; add new methods to StaffService for CRUD operations and filtering
